django_rest_framework/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
